+++ tests/integrate/display.rs
